generate operand groups
